pyproject.toml
[project]
name = "video-transcription-analysis"
version = "0.1.0"
description = "Local video transcription and analysis using Ollama"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "mlx-whisper>=0.4.3",
    "torch>=2.0.0",
    "ffmpeg-python>=0.2.0",
    "pyannote.audio>=3.0.0",
    "speechbrain>=1.0.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_transcription"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
